outout control
