--- client/styles/_overwrites.scss
@@ -1,66 +1,0 @@
-.publisherManager.modal {
-    label {
-        min-width: auto;
-    }
-    .modal__header {
-        background-color: white;
-    }
-}
-
-.navigation-main {
-    .subnav {
-        top: 0;
-    }
-    .flex-grid {
-        position: absolute;
-        top: 48px;
-    }
-}
-
-.publisher-mediaBox {
-    overflow: visible !important;
-    .toolbox {
-        overflow: visible !important;
-        z-index: 1 !important;
-        visibility: hidden;
-
-        .dropdown {
-            line-height: 5.6rem;
-            position: absolute;
-            top: 0;
-            bottom: 0;
-            left: 0;
-            right: 0;
-            background: linear-gradient(to right, rgba(244, 244, 244, 0) 0%, rgba(244, 244, 244, 0) 0%, #f4f4f4 40%);
-            text-align: center;
-        }
-
-        .dropdown__menu {
-            z-index: 2 !important;
-
-        }
-        * {
-            overflow: visible !important;
-        }
-        .dropdown__toggle {
-            visibility: hidden;
-        }
-    }
-
-    &:hover {
-        .toolbox{
-            visibility: visible;
-        }
-        .toolbox .dropdown__toggle {
-            visibility: visible;
-        }
-    }
-}
-.flex-item {
-    display: flex;
-    flex-direction: column;
-}
-
-.compact-view .sp-list-item-view .media-box .item-info .state-label {
-    overflow: visible;
-}
