+++ client/styles/_publisher.scss
@@ -4,206 +4,11 @@
 
 @import '~mixins.scss';
 @import '~variables.scss';
-
-// @import '_overwrites.scss';
-// @import 'sidebarLayout.scss';
-// @import 'nav.scss';
-// @import 'paneLayout.scss';
 @import 'treeList.scss';
-// @import 'flatList.scss';
 @import 'listElement.scss';
-// @import 'helperElements.scss';
-// @import 'paneContent.scss';
-// @import 'sectionTreeElement.scss';
-// @import 'verticalLayout.scss';
+@import 'helperElements.scss';
 @import 'dragNdrop.scss';
-// @import 'dropdownLight.scss';
-// @import 'flexBox.scss';
-// @import 'darkFilters.scss';
-// @import 'checkButton.scss';
-// @import 'articlePreview.scss';
-// @import 'removableLabels.scss';
+@import 'articlePreview.scss';
 @import 'wizardThemeBox.scss';
 @import 'wizardThemeDetails.scss';
 @import 'iconOnHover.scss';
-// @import 'numberLabel.scss';
-
-
-.sp-logo-wordpress {
-    display: block;
-    float: right;
-    margin-right: 1rem;
-    width: 30px;
-    height: 30px;
-    background: transparent url(../images/wordpress-icon.svg) center center no-repeat;
-}
-
-.absPosition {
-    position: absolute;
-
-    &--bottomRight {
-        bottom: 1rem;
-        right: 1rem;
-    }
-
-    &--top {
-        top: 0;
-    }
-}
-
-.margin {
-    &--none {
-        margin: 0!important;
-    }
-
-    &--top {
-        margin-top: 1em;
-    }
-
-    &--top0 {
-        margin-top: 0 !important;
-    }
-
-    &--topBig {
-        margin-top: 2rem!important;
-    }
-
-    &--bottom {
-        margin-bottom: 1em;
-    }
-
-    &--bottomHalf {
-        margin-bottom: .5em;
-    }
-
-    &--bottom0 {
-        margin-bottom: 0;
-    }
-
-    &--right {
-        margin-right: 1em;
-    }
-}
-
-.padding {
-    &--right0 {
-        padding-right: 0;
-    }
-    &--bottomBig {
-        padding-bottom: 4rem !important;
-    }
-}
-
-.text {
-    &--white {
-        color: $white !important;
-    }
-    &--light {
-        color: $gray;
-    }
-
-    &--lighter {
-        color: $grayLight;
-    }
-
-    &--dark {
-        color: $grayDarker;
-    }
-
-    &--green {
-        color: $green;
-    }
-
-    &--red {
-        color: $red;
-    }
-
-    &--italic {
-        font-style: italic;
-    }
-
-    &--bold {
-        font-weight: 500;
-    }
-
-    &--bolder {
-        font-weight: 700;
-    }
-
-    &--thin {
-        font-weight: 300;
-    }
-
-    &--normal {
-        font-weight: 400;
-    }
-
-    &--uppercase {
-        text-transform: uppercase;
-    }
-
-    &--small {
-        font-size: .9em;
-    }
-
-    &--smallish {
-        font-size: .8em;
-    }
-
-    &--smaller {
-        font-size: 9px;
-    }
-
-    &--smallHeading {
-        font-size: 12px;
-        line-height: 1em;
-        margin-bottom: .8em;
-        color: $gray;
-    }
-
-    &--strikethrough {
-        text-decoration: line-through;
-    }
-
-    &--verticalMiddle {
-        vertical-align: middle;
-    }
-
-    &--blue {
-        color: $sd-blue;
-    }
-
-    &--keyword {
-        color: $sd-keyword;
-    }
-
-    &--center {
-        text-align: center;
-    }
-
-    &--left {
-        text-align: left !important;
-    }
-
-    &--right {
-        text-align: right !important;
-    }
-}
-
-.inlineBlock {
-    display: inline-block;
-}
-
-.iconButton {
-    border: none;
-    background: none;
-    padding: 0;
-}
-
-.relative {
-    position: relative;
-}
-
-.overflow-auto {
-    overflow: auto;
-}
