--- client/styles/dropdownLight.scss
@@ -1,21 +1,0 @@
-.dropdownLight {
-    .dropdown {
-        display: block;
-    }
-
-    .dropdown__toggle.line-input {
-        border-color: rgba(255,255,255, .15) !important;
-        color: #eaeaea;
-        line-height: 20px;
-        height: 20px;
-        vertical-align: baseline;
-    }
-
-    .dropdown__caret {
-        border-top-color: #fff;
-        position: absolute;
-        top: 50%;
-        right: 3px;
-        margin-top: -3px;
-    }
-}
