--- client/styles/flatList.scss
@@ -1,3 +1,0 @@
-.flatList {
-	border-top: 1px solid $tableBorder;
-}
