--- client/styles/numberLabel.scss
@@ -1,18 +1,0 @@
-.numberLabel {
-    display: flex;
-    &__number {
-        color: $grayLight;
-        font-size: 3rem;
-        line-height: 3rem;
-        font-weight: 200;
-        margin-right: 0.4rem;
-    }
-
-    &__label {
-        color: $grayLight;
-        text-transform: uppercase;
-        font-size: 1rem;
-        line-height: 1.3rem;
-        padding-top: .1rem;
-    }
-}
