--- client/styles/verticalLayout.scss
@@ -1,20 +1,0 @@
-.verticalLayout {
-    display: flex;
-    flex-direction: column;
-
-    &--directionRow {
-        flex-direction: row;
-    }
-
-    &__element {
-        position: relative;
-        flex-grow: 1;
-        flex-basis: 50%;
-        overflow-x: hidden;
-        overflow-y: scroll;
-
-        &--border {
-            border-bottom: 2px solid $tableBorder;
-        }
-    }
-}
