--- client/views/monitoring/index.html
@@ -1,66 +1,0 @@
-<div class="subnav  subnav--absolute">
-    <div class="flat-searchbar monitoring-backlink" ng-if="webPublisherMonitoring.tenantArticles">
-        <div class="search-handler">
-            <button tooltip="{{ :: 'Back to previous view' | translate }}"
-                tooltip-placement="right"
-                class="backlink"
-                ng-click="webPublisherMonitoring.viewMonitoringHome()"></button>
-        </div>
-    </div>
-    <div class="filter-trigger"
-        title="Filters"
-        ng-click="webPublisherMonitoring.toggleFilterPane()"
-        ng-class="{'filter-trigger--active': webPublisherMonitoring.filterOpen}">
-            <i class="icon-filter-large"></i>
-    </div>
-    <div class="flat-searchbar monitoring-flat-searchbar"
-        ng-class="{extended: publisher.monitoringSearch || webPublisherMonitoring.advancedFilters.term}">
-        <div class="search-handler">
-            <label for="search-input"
-                class="trigger-icon"
-                ng-click="publisher.monitoringSearch = !publisher.monitoringSearch">
-                <i class="icon-search"></i>
-            </label>
-            <input id="search-input"
-                type="text"
-                placeholder="{{ :: 'Search' | translate }}"
-                ng-model="webPublisherMonitoring.advancedFilters.term"
-                ng-model-options="{debounce: 1000}">
-            <button class="search-close"
-                ng-click="webPublisherMonitoring.advancedFilters.term = ''"
-                ng-class="{visible: webPublisherMonitoring.advancedFilters.term}">
-                <i class="icon-remove-sign"></i>
-            </button>
-       </div>
-    </div>
-    <h3 class="page-nav-title">Output Control</h3>
-</div>
-
-<div class="paneLayout paneLayout--twoHeadersMargin">
-    <div class="paneLayout__content">
-        <!-- paneLayout inside paneLayout :) -->
-        <div class="paneLayout">
-            <div class="paneLayout__leftPane paneContent"
-                ng-class="{'paneLayout__leftPane--open': webPublisherMonitoring.filterOpen}"
-                ng-include="'filter-pane.html'">
-            </div>
-
-            <div class="paneLayout__content">
-                <div class="contentPadding">
-                    <div ng-include="'filter-labels.html'"></div>
-                    <div ng-include="'index-content.html'"></div>
-                </div>
-            </div>
-        </div>
-    </div>
-    <div class="paneLayout__pane paneContent paneContent--darker"
-        ng-show="webPublisherMonitoring.publishOpen"
-        ng-class="{'paneLayout__pane--open': webPublisherMonitoring.publishOpen}"
-        ng-include="'publishing-pane.html'">
-    </div>
-    <div class="paneLayout__pane"
-        ng-show="webPublisherMonitoring.previewOpen"
-        ng-class="{'paneLayout__pane--open': webPublisherMonitoring.previewOpen}"
-        ng-include="'preview-pane.html'">
-    </div>
-</div>
