--- client/views/monitoring/publish-pane-listitem.html
@@ -1,81 +1,0 @@
-<div ng-show="webPublisherMonitoring.publishFilter == 'all' || webPublisherMonitoring.publishedDestinations[site.code].status == webPublisherMonitoring.publishFilter"
-    class="listElement listElement--dark listElement--noBorder"
-    ng-class="{'listElement--borderPublished listElement--lightDark': webPublisherMonitoring.publishedDestinations[site.code].status == 'published',
-    'listElement--borderUnpublished': webPublisherMonitoring.publishedDestinations[site.code].status == 'unpublished'}">
-    <div class="listElement__content listElement__content--dark">
-        <div class="minHeightLine">
-            {{site.name}}
-        </div>
-        <div ng-if="!site.outputChannel" class="text--uppercase text--smallish text--bold text--lighter">route</div>
-        <div>
-            <div ng-if="!webPublisherMonitoring.publishedDestinations[site.code]" class="flexBox">
-                <div ng-if="!site.outputChannel"
-                    sd-publish-routes data-site="site"
-                    data-monitoring-contoller="webPublisherMonitoring"
-                    class="flexBox__fullItem" style="margin-top: -1.8rem;">
-                </div>
-                <sd-check
-                    ng-if="site.outputChannel"
-                    ng-model="webPublisherMonitoring.newDestinations[site.code].status"
-                    ng-true-value="published" ng-false-value="new"></sd-check>
-                <span ng-if="site.outputChannel" class="text--uppercase text--smallish text--bold text--lighter">publish</span>
-
-            </div>
-            <div ng-if="webPublisherMonitoring.publishedDestinations[site.code] && webPublisherMonitoring.publishedDestinations[site.code].status != 'unpublished'" >
-                <div ng-show="site.routeEditing" class="flexBox">
-                    <div sd-publish-routes data-site="site" data-monitoring-contoller="webPublisherMonitoring" class="flexBox__fullItem" style="margin-top: -1.8rem;"></div>
-                    <div class="flexBox__item">
-                        <i class="icon-close-small icon--white text--verticalMiddle cursorPointer rightSpace"
-                            ng-click="site.routeEditing = false"></i>
-                    </div>
-                </div>
-                <span
-                    class="label text--smaller text--uppercase" ng-class="webPublisherMonitoring.publishedDestinations[site.code].status == 'new' ? 'label--info' : 'label--success'"
-                    ng-hide="site.routeEditing">{{webPublisherMonitoring.publishedDestinations[site.code].route.name}}
-                </span>
-                <span class="text--uppercase text--smaller text--blue cursorPointer leftSpace"
-                    ng-click="site.routeEditing = true"
-                    ng-hide="site.routeEditing">
-                    change
-                </span>
-            </div>
-            <span ng-if="webPublisherMonitoring.publishedDestinations[site.code].status == 'unpublished'" class="label text--smaller text--uppercase">unpublished</span>
-            <span ng-if="webPublisherMonitoring.publishedDestinations[site.code].status == 'unpublished'"
-                class="text--uppercase text--smaller text--blue cursorPointer leftSpace"
-                ng-click="webPublisherMonitoring.publishedDestinations[site.code].status = null"
-                >Publish</span>
-        </div>
-    </div>
-    <div class="listElement__right">
-        <div class="text--small text--lighter text--right minHeightLine">
-            <span ng-if="webPublisherMonitoring.newDestinations[site.code].route.id && webPublisherMonitoring.publishedDestinations[site.code].route.id == webPublisherMonitoring.newDestinations[site.code].route.id">
-                {{webPublisherMonitoring.publishedDestinations[site.code].updatedAt | reldate}}
-            </span>
-            <a ng-if='webPublisherMonitoring.newDestinations[site.code].liveUrl && webPublisherMonitoring.publishedDestinations[site.code].route.id == webPublisherMonitoring.newDestinations[site.code].route.id'
-                class="text--uppercase text--small text--blue cursorPointer rightSpace--small"
-                ng-href="{{webPublisherMonitoring.newDestinations[site.code].liveUrl}}"
-                target="_blank">
-                <span sd-tooltip="Live Preview"><i class="icon-external icon--blue text--verticalMiddle rightSpace--small"></i></span>
-            </a>
-            <span ng-if='webPublisherMonitoring.newDestinations[site.code].route.id && ( !webPublisherMonitoring.newDestinations[site.code].liveUrl || webPublisherMonitoring.publishedDestinations[site.code].route.id != webPublisherMonitoring.newDestinations[site.code].route.id) && webPublisherMonitoring.newDestinations[site.code].status != "unpublished"'
-                class="text--uppercase text--small text--blue cursorPointer rightSpace--small"
-                ng-click="webPublisherMonitoring.openArticlePreview(webPublisherMonitoring.newDestinations[site.code].route.id, site)">
-                <span sd-tooltip="Live Preview"><i class="icon-external icon--blue text--verticalMiddle rightSpace--small"></i></span>
-            </span>
-        </div>
-        <div class="minHeightLine text--right">
-            <span ng-if="webPublisherMonitoring.publishedDestinations[site.code].fbia" class="label label--facebook text--smaller text--uppercase">facebook</span>
-            <div ng-if="webPublisherMonitoring.newDestinations[site.code].route.id && !webPublisherMonitoring.publishedDestinations[site.code].fbia && webPublisherMonitoring.publishedDestinations[site.code].status != 'unpublished'">
-                <span class="text--uppercase text--smallish text--bold text--lighter">publish to facebook</span>
-                <sd-check ng-model="webPublisherMonitoring.newDestinations[site.code].fbia"></sd-check>
-            </div>
-        </div>
-        <div class="minHeightLine text--right">
-            <span ng-if="webPublisherMonitoring.publishedDestinations[site.code].fbia" class="label label--facebook text--smaller text--uppercase">facebook</span>
-            <div ng-if="webPublisherMonitoring.newDestinations[site.code].route.id && !webPublisherMonitoring.publishedDestinations[site.code].fbia && webPublisherMonitoring.publishedDestinations[site.code].status != 'unpublished'">
-                <span class="text--uppercase text--smallish text--bold text--lighter">Paywall Secured</span>
-                <sd-check ng-model="webPublisherMonitoring.newDestinations[site.code].paywallSecured"></sd-check>
-            </div>
-        </div>
-    </div>
-</div>
