--- client/views/monitoring/sections-treeElement.html
@@ -1,17 +1,0 @@
-<div class="sectionTreeElement">
-    <div class="sectionTreeElement__icon" >
-        <sd-check ng-model="webPublisherMonitoring.advancedFilters.sites[site.code]['routes'][route.id].status"
-            ng-click="webPublisherMonitoring.advancedFilters.sites[site.code]['routes'][route.id].name = route.name; webPublisherMonitoring.advancedFilters.sites[site.code].name = site.name">
-        </sd-check>
-    </div>
-    <button class="sectionTreeElement__icon iconButton" ng-class="{'sectionTreeElement__icon--extraMargin': route.children.length}" ng-if="route.children.length">
-        <i class="icon-chevron-down-thin" ng-show="route.filterPaneOpen" ng-click="route.filterPaneOpen = !route.filterPaneOpen"></i>
-        <i class="icon-chevron-right-thin" ng-show="!route.filterPaneOpen" ng-click="route.filterPaneOpen = !route.filterPaneOpen"></i>
-    </button>
-    <div class="sectionTreeElement__content">
-        {{route.name}} <i class="sectionTreeElement__count">{{route.articlesCount}}</i>
-    </div>
-</div>
-<ul ng-show="route.children.length && route.filterPaneOpen">
-    <li ng-repeat="route in route.children" ng-include="'sections-treeElement.html'"></li>
-</ul>
