+++ client/views/output/article-preview.html
@@ -1,0 +1,42 @@
+<div sd-modal data-model="true" data-full="true" class="modal--fullscreen publisherManager">
+    <div class="modal__header">
+        <div class="flexBox">
+            <h6 class="flexBox__item text--left width200">Live preview: <b class="text--dark">{{webPublisherOutput.articlePreviewModeReadable}}</b></h6>
+            <div class="flexBox__fullItem flexBox__fullItem--margin text--center">
+                <a class="cursorPointer previewIcon previewIcon--desktop"
+                    ng-class="{'previewIcon--active': webPublisherOutput.articlePreviewMode == 'desktop'}"
+                    ng-click="webPublisherOutput.setArticlePreviewMode('desktop')"></a>
+                <a class="cursorPointer previewIcon previewIcon--tablet"
+                    ng-class="{'previewIcon--active': webPublisherOutput.articlePreviewMode == 'tablet'}"
+                    ng-click="webPublisherOutput.setArticlePreviewMode('tablet')"></a>
+                <a class="cursorPointer previewIcon previewIcon--tabletLandscape"
+                    ng-class="{'previewIcon--active': webPublisherOutput.articlePreviewMode == 'tablet-landscape'}"
+                    ng-click="webPublisherOutput.setArticlePreviewMode('tablet-landscape')"></a>
+                <a class="cursorPointer previewIcon previewIcon--mobile"
+                    ng-class="{'previewIcon--active': webPublisherOutput.articlePreviewMode == 'mobile'}"
+                    ng-click="webPublisherOutput.setArticlePreviewMode('mobile')"></a>
+                <a class="cursorPointer previewIcon previewIcon--mobileLandscape"
+                    ng-class="{'previewIcon--active': webPublisherOutput.articlePreviewMode == 'mobile-landscape'}"
+                    ng-click="webPublisherOutput.setArticlePreviewMode('mobile-landscape')"></a>
+                <a class="cursorPointer previewIcon previewIcon--amp"
+                    ng-class="{'previewIcon--active': webPublisherOutput.articlePreviewMode == 'amp'}"
+                    ng-click="webPublisherOutput.setArticlePreviewMode('amp')"></a>
+                <a class="cursorPointer previewIcon previewIcon--ampLandscape"
+                    ng-class="{'previewIcon--active': webPublisherOutput.articlePreviewMode == 'amp-landscape'}"
+                    ng-click="webPublisherOutput.setArticlePreviewMode('amp-landscape')"></a>
+            </div>
+            <div class="flexBox__item text--right width200">
+                <button class="btn" ng-click="webPublisherOutput.openArticlePreviewModal = false" translate>Close</button>
+            </div>
+        </div>
+    </div>
+    <div class="modal__body articlePreview">
+        <iframe ng-src="{{webPublisherOutput.previewArticleSrc}}"
+            class="articlePreview__iframe"
+            ng-class="{'articlePreview__iframe--tablet': webPublisherOutput.articlePreviewMode == 'tablet',
+                        'articlePreview__iframe--tabletLandscape': webPublisherOutput.articlePreviewMode == 'tablet-landscape',
+                        'articlePreview__iframe--mobile': webPublisherOutput.articlePreviewMode == 'mobile' || webPublisherOutput.articlePreviewMode == 'amp',
+                        'articlePreview__iframe--mobileLandscape': webPublisherOutput.articlePreviewMode == 'mobile-landscape' || webPublisherOutput.articlePreviewMode == 'amp-landscape'}">
+        </iframe>
+    </div>
+</div>
