+++ client/views/output/filter-pane.html
@@ -1,0 +1,56 @@
+<div class="side-panel side-panel--transparent side-panel--shadow-right">
+    <div class="side-panel__header side-panel__header--border-b">
+            <a class="icn-btn side-panel__close" sd-tooltip="Close filters" flow="left" ng-click="webPublisherOutput.toggleFilterPane()"><i class="icon-close-small"></i></a>
+        <h3 class="side-panel__heading side-panel__heading--big">Advanced Filter</h3>
+    </div>
+    <div class="side-panel__content">
+        <div class="sd-loader" ng-if="loading"></div>
+        <div class="side-panel__content-block">
+            <div class="form__row">
+                <div class="sd-line-input sd-line-input--no-margin">
+                    <sd-tag-input
+                        ng-if="webPublisherOutput.routes"
+                        ng-model="webPublisherOutput.advancedFilters.routes"
+                        data-label="Routes"
+                        data-items="webPublisherOutput.routes"
+                        data-field="name"
+                        data-freetext="false" >
+                        <tags-input
+                            ng-model="model"
+                            add-from-autocomplete-only="{{items ? !freetext : false}}"
+                            show-button="{{!!items}}"
+                            display-property="{{field}}"
+                            use-strings="{{!field}}"
+                        ></tags-input>
+                    </sd-tag-input>
+                </div>
+            </div>
+            <div class="form__row">
+                <div class="sd-line-input sd-line-input--no-margin">
+                    <sd-tag-input ng-model="webPublisherOutput.advancedFilters.author" data-label="Authors">
+                        <tags-input ng-model="model" add-from-autocomplete-only="{{items ? !freetext : false}}" show-button="{{!!items}}" display-property="{{field}}" use-strings="{{!field}}"></tags-input>
+                    </sd-tag-input>
+                </div>
+            </div>
+            <div class="form__row form__row--flex">
+                <div class="sd-line-input sd-line-input--no-margin">
+                    <label class="sd-line-input__label">Published after</label>
+                    <div sd-datepicker data-format="YYYY-MM-DD" name="publishedAfter" id="publishedAfter" ng-model="webPublisherOutput.advancedFilters.publishedAfter"></div>
+                </div>
+            </div>
+            <div class="form__row form__row--flex">
+                <div class="sd-line-input sd-line-input--no-margin">
+                    <label class="sd-line-input__label">Published before</label>
+                    <div sd-datepicker data-format="YYYY-MM-DD" name="publishedBefore" id="publishedBefore" ng-model="webPublisherOutput.advancedFilters.publishedBefore"></div>
+                </div>
+            </div>
+            <div class="form__row">
+                <div class="sd-line-input sd-line-input--no-margin">
+                    <sd-tag-input ng-model="webPublisherOutput.advancedFilters.source" data-label="Ingest Source">
+                        <tags-input ng-model="model" add-from-autocomplete-only="{{items ? !freetext : false}}" show-button="{{!!items}}" display-property="{{field}}" use-strings="{{!field}}"></tags-input>
+                    </sd-tag-input>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
